Fix CI failures from strict warning flags across all compilers
The new -Wshadow=local, -Wnon-virtual-dtor, -Wdouble-promotion, and
-Wformat=2 flags exposed warnings-as-errors across many compiler and
toolkit combinations. This commit fixes them all:

CMake (CCCLBuildCompilerTargets.cmake):
- Remove -Wnull-dereference: GCC produces false positives in system
  headers (e.g. streambuf) when inlining user code.
- Guard -Wgnu / -Wno-gnu-* flags behind a Clang-only check: GCC
  (especially GCC9) rejects unrecognized -Wno-* flags under -Werror.

-Wdouble-promotion fixes (float-to-double implicit promotion):
- libcudacxx complex<double> constructors/assignments from __half and
  __nv_bfloat16 in nvfp16.h and nvbf16.h: wrap __half2float() and
  __bfloat162float() results in static_cast<double>().
- libcudacxx native_type.h: use __float128{} instead of __float128{0.0}
  to avoid double-to-__float128 promotion.
- thrust complex arithmetic.h: cast all mixed-type operands to their
  common_type before arithmetic to make promotion explicit.
- thrust complex.inl equality operators: same common_type cast pattern.
- thrust catch2_test_complex.cu: cast to double before CHECK comparisons
  of mixed float/double complex components.
- nvbench_helper.cu: fix float-to-double promotion in random_to_item_t
  calls and uniform_real_distribution<double> constructor literals.
- gen_uniform_distribution.cu: explicit casts in mixed-type arithmetic.
- cudax 09-nbody-blocked.cu: cast float elapsed to double for printf.

-Wshadow=local fixes (local variable shadowing):
- nvbench_helper.cu: rename dist -> float_dist / double_dist to avoid
  shadowing the function parameter.
- gen_entropy.cu: rename lambda parameter entropy -> level.
- cudax sections_2.cu: rename inner s_bar -> s_baz.

-Wnon-virtual-dtor fix:
- catch2_test_helper.h: change ~QuietMatchExpr() override to virtual
  ~QuietMatchExpr() since ITransientExpression has no virtual dtor.

+++ thrust/testing/catch2_test_complex.cu
@@ -213,8 +213,8 @@ TEMPLATE_LIST_TEST_CASE("ComplexConstructionAndAssignmentWithPromoting", "[compl
     thrust::complex<T0> copy_assign{};
     const thrust::complex<T1> expected{real_T1, imag_T1};
     copy_assign = expected;
-    CHECK(expected.real() == copy_assign.real());
-    CHECK(expected.imag() == copy_assign.imag());
+    CHECK(static_cast<double>(expected.real()) == static_cast<double>(copy_assign.real()));
+    CHECK(static_cast<double>(expected.imag()) == static_cast<double>(copy_assign.imag()));
   }
 
   {
@@ -223,7 +223,7 @@ TEMPLATE_LIST_TEST_CASE("ComplexConstructionAndAssignmentWithPromoting", "[compl
     const T1 to_be_copied = real_T1;
     assign_from_lvalue_T  = to_be_copied;
     require_almost_equal(expected.real(), assign_from_lvalue_T.real());
-    CHECK(expected.imag() == assign_from_lvalue_T.imag());
+    CHECK(static_cast<double>(expected.imag()) == static_cast<double>(assign_from_lvalue_T.imag()));
   }
 
   {
